refactor(test): #320 split entity_manager_update_batch_test.go by concern
Update tests to entity_manager_update_test.go, batch tests to
entity_batch_service_test.go, shared visitPayload fixture joins
entity_manager_mocks_test.go. Pure mechanical move; inventory unchanged.

Refs #320 #271 #401
